Log cleanup

+++ src/SongCacheIndex.cpp
@@ -511,7 +511,7 @@ SongCacheIndex::CacheSong(Song& song, string dir)
 		insertSong.bind(index++, GetHashForDirectory(song.GetSongDir()));
 		insertSong.bind(index++, song.GetSongDir());
 		insertSong.exec();
-		int songID = sqlite3_last_insert_rowid(db->getHandle());
+		int64_t songID = sqlite3_last_insert_rowid(db->getHandle());
 		vector<Steps*> vpStepsToSave = song.GetStepsToSave();
 		FOREACH_CONST(Steps*, vpStepsToSave, s)
 		{
@@ -523,7 +523,7 @@ SongCacheIndex::CacheSong(Song& song, string dir)
 						  dir.c_str());
 				continue;
 			}
-			int stepsID = InsertSteps(pSteps, songID);
+			int64_t stepsID = InsertSteps(pSteps, songID);
 		}
 		return true;
 	} catch (std::exception& e) {
@@ -887,7 +887,7 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 	SSCLoader loader;
 	int songid = query.getColumn(0);
 	int index = 1;
-	song->m_fVersion = static_cast<double>(query.getColumn(index++));
+	song->m_fVersion = static_cast<float>(static_cast<double>(query.getColumn(index++)));
 	song->m_sMainTitle = static_cast<const char*>(query.getColumn(index++));
 	song->m_sSubTitle = static_cast<const char*>(query.getColumn(index++));
 	song->m_sArtist = static_cast<const char*>(query.getColumn(index++));
@@ -916,11 +916,11 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 	loader.ProcessInstrumentTracks(
 	  *song, static_cast<const char*>(query.getColumn(index++)));
 	song->m_SongTiming.m_fBeat0OffsetInSeconds =
-	  static_cast<double>(query.getColumn(index++));
+		static_cast<float>(static_cast<double>(query.getColumn(index++)));
 	song->m_fMusicSampleStartSeconds =
-	  static_cast<double>(query.getColumn(index++));
+		static_cast<float>(static_cast<double>(query.getColumn(index++)));
 	song->m_fMusicSampleLengthSeconds =
-	  static_cast<double>(query.getColumn(index++));
+		static_cast<float>(static_cast<double>(query.getColumn(index++)));
 
 	int selection = static_cast<int>(query.getColumn(index++));
 	if (selection == 0)
@@ -930,8 +930,8 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 
 	int bpmminIndex = index++;
 	int bpmmaxIndex = index++;
-	float BPMmin = static_cast<double>(query.getColumn(bpmminIndex));
-	float BPMmax = static_cast<double>(query.getColumn(bpmmaxIndex));
+	float BPMmin = static_cast<float>(static_cast<double>(query.getColumn(bpmminIndex)));
+	float BPMmax = static_cast<float>(static_cast<double>(query.getColumn(bpmmaxIndex)));
 	if (query.isColumnNull(bpmminIndex) || query.isColumnNull(bpmmaxIndex)) {
 		if (query.isColumnNull(bpmminIndex) && query.isColumnNull(bpmmaxIndex))
 			song->m_DisplayBPMType = DISPLAY_BPM_ACTUAL;
@@ -973,7 +973,7 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 						 static_cast<const char*>(query.getColumn(index++)),
 						 title);
 
-	song->SetSpecifiedLastSecond(static_cast<double>(query.getColumn(index++)));
+	song->SetSpecifiedLastSecond(static_cast<float>(static_cast<double>(query.getColumn(index++))));
 
 	string animations = static_cast<const char*>(query.getColumn(index++));
 	string animationstwo = static_cast<const char*>(query.getColumn(index++));
@@ -994,12 +994,12 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 		keysounds = keysounds.substr(1);
 	}
 	split(keysounds, ",", song->m_vsKeysoundFile);
-	song->SetFirstSecond(static_cast<double>(query.getColumn(index++)));
-	song->SetLastSecond(static_cast<double>(query.getColumn(index++)));
+	song->SetFirstSecond(static_cast<float>(static_cast<double>(query.getColumn(index++))));
+	song->SetLastSecond(static_cast<float>(static_cast<double>(query.getColumn(index++))));
 	song->m_sSongFileName = static_cast<const char*>(query.getColumn(index++));
 	song->m_bHasMusic = static_cast<int>(query.getColumn(index++)) != 0;
 	song->m_bHasBanner = static_cast<int>(query.getColumn(index++)) != 0;
-	song->m_fMusicLengthSeconds = static_cast<double>(query.getColumn(index++));
+	song->m_fMusicLengthSeconds = static_cast<float>(static_cast<double>(query.getColumn(index++)));
 	int dirhash = query.getColumn(index++);
 	string dir = query.getColumn(index++);
 	song->SetSongDir(dir);
@@ -1068,7 +1068,7 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 				  TimingData(song->m_SongTiming.m_fBeat0OffsetInSeconds);
 				// Load timing data
 				stepsTiming.m_fBeat0OffsetInSeconds =
-				  static_cast<double>(qTiming.getColumn(timingIndex++));
+					static_cast<float>(static_cast<double>(qTiming.getColumn(timingIndex++)));
 				SSCLoader::ProcessBPMs(
 				  stepsTiming,
 				  static_cast<const char*>(qTiming.getColumn(timingIndex++)),
@@ -1132,8 +1132,8 @@ SongCacheIndex::SongFromStatement(Song* song, SQLite::Statement& query)
 
 		int bpmminIndex = stepsIndex++;
 		int bpmmaxIndex = stepsIndex++;
-		float BPMmin = static_cast<double>(qSteps.getColumn(bpmminIndex));
-		float BPMmax = static_cast<double>(qSteps.getColumn(bpmmaxIndex));
+		float BPMmin = static_cast<float>(static_cast<double>(qSteps.getColumn(bpmminIndex)));
+		float BPMmax = static_cast<float>(static_cast<double>(qSteps.getColumn(bpmmaxIndex)));
 		if (qSteps.isColumnNull(bpmminIndex) ||
 			qSteps.isColumnNull(bpmmaxIndex)) {
 			if (qSteps.isColumnNull(bpmminIndex) &&
